Don't delete namespaces in terminating phase

+++ pkg/custregion/gc.go
@@ -4,6 +4,7 @@ import (
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"github.com/sirupsen/logrus"
+	"k8s.io/api/core/v1"
 )
 
 func Collect(kubeApi kubernetes.Interface,
@@ -24,6 +25,11 @@ func Collect(kubeApi kubernetes.Interface,
 	log.Infof("there are %d namespaces", len(nses.Items))
 	for _, ns := range nses.Items {
 		if !isKnownCustRegion(ns.Name) {
+			if ns.Status.Phase == v1.NamespaceTerminating {
+				log.Debugf("skipping deletion of terminating namespace %s",
+					ns.Name)
+				continue
+			}
 			log.Infof("deleting orphaned namespace %s", ns.Name)
 			err = nsApi.Delete(ns.Name, nil)
 			if err != nil {
